requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
pandas==3.0.2
sqlalchemy==2.0.49 
pyodbc==5.3.0
requests==2.33.1